% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePerc.R
\name{simplePerc}
\alias{simplePerc}
\title{Simple Perceptron}
\usage{
simplePerc(X, yt, act.fn = "step", niter = 20, lrate = 0.6)
}
\arguments{
\item{X}{Data frame with numeric columns}

\item{yt}{Training labels (numeric)}

\item{act.fn}{Activation function. Accepts "step","relu", "sigmoid", "tanh" and "gauss". Step is the default value.}

\item{niter}{Number of iterations. Default is 20}

\item{lrate}{Learning rate. Default is 0.6}
}
\value{
A list of errors at each iteration in classification
}
\description{
A single layer perceptron
}
\examples{
err = simplePerc(X,yt,"step",30,0.4)
}
\seealso{
\link[mlbench]{Sonar}
Sonar data set
}
